Reorganized new preferences files
