+++ src/app/newSettings/Sections/Account/Security/components/EnterPassword.tsx
