+++ src/app/newSettings/Sections/Account/Security/components/ExportBackupKey.tsx
