+++ src/app/newSettings/Sections/Workspace/Members/components/ActivityTab.tsx
@@ -1,4 +1,4 @@
-import { ActivityTabProps } from '../../../../types';
+import { ActivityTabProps } from '../../../../types/types';
 
 const ActivityTab = ({ user, activity }: ActivityTabProps): JSX.Element => {
   return <div>ACTIVITY TAB</div>;
