+++ src/app/newSettings/Sections/Workspace/Members/components/TeamsTab.tsx
@@ -1,4 +1,4 @@
-import { TeamsTabProps } from '../../../../types';
+import { TeamsTabProps } from '../../../../types/types';
 
 const TeamsTab = ({ user }: TeamsTabProps): JSX.Element => {
   return <div>TEAMS TAB</div>;
