+++ src/app/newSettings/Sections/Workspace/Members/components/UserCard.tsx
@@ -1,6 +1,6 @@
 import { useTranslationContext } from '../../../../../i18n/provider/TranslationProvider';
 import Avatar from '../../../../../shared/components/Avatar';
-import { MemberRole } from '../../../../types';
+import { MemberRole } from '../../../../types/types';
 import RoleBadge from './RoleBadge';
 
 interface UserProps {
